% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.r
\name{dmrff.annotate}
\alias{dmrff.annotate}
\title{dmrff.annotate}
\usage{
dmrff.annotate(regions, annotation)
}
\arguments{
\item{regions}{Data frame listing the regions to annotate.
Must have columns "chr", "start" and "end" to specify genomic regions.}

\item{annotation}{Data frame listing the features that will be used to
annotate the genomic regions.
Must have columns "chr" and "pos" to identify genomic locations.}
}
\value{
A data frame list all features contained in the regions along
with the regions they belong to.
}
\description{
Annotate a set of regions with feature annotations.
}
\details{
For example, the regions could be differentially methylated regions
and the features could be CpG sites.
}
\examples{

methylation <- ... ## methylation matrix
ewas <- data.frame(chr=..., ## chromosome of each CpG site
                   pos=..., ## chromosomal position of each CpG site
                   symbol=..., ## symbol of gene linked to each CpG site
                   ..., ## other CpG site annotations
                   estimate=..., ## EWAS effect estimate
                   se=...,       ## EWAS standard error of the estimate 
                   p.value=...)  ## EWAS p-value
                
dmrs <- with(ewas, dmrff(estimate, se, p.value, methylation, chr, pos))
dmrs <- dmrs[which(dmrs$p.adjust < 0.05),]
dmrs <- dmrff.annotate(dmrs, annotation)

}
