% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.r
\name{dmrff.candidates}
\alias{dmrff.candidates}
\title{dmrff.candidates}
\usage{
dmrff.candidates(estimate, p.value, chr, pos, maxgap = 500,
  p.cutoff = 0.05, verbose = T)
}
\arguments{
\item{estimate}{Vector of EWAS effect estimates
(corresponds to rows of \code{methylation}).}

\item{p.value}{Vector of p-values.}

\item{chr}{Feature chromosome (corresponds to rows of \code{methylation}).}

\item{pos}{Feature chromosome position.}

\item{maxgap}{Maximum distance between consecutive features (Default: 500bp).}

\item{p.cutoff}{Unadjusted p-value cutoff for membership in a candidate DMR
(Default: 0.05).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}
}
\value{
A data frame listing all candidate regions.
}
\description{
Identify candidate differentially methylated regions from EWAS summary statistics.
}
