% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{dmrff.manhattan.plot}
\alias{dmrff.manhattan.plot}
\title{Manhattan plot}
\usage{
dmrff.manhattan.plot(object, chr, pos, title = "Manhattan plot")
}
\arguments{
\item{object}{Output from the \code{\link{dmrff}()}, \code{\link{dmrff.meta}()}
or \code{\link{dmrff.cohort}} functions.}

\item{title}{Title for the plot (Default: "Manhattan plot").}
}
\value{
\code{\link{ggplot}} object showing the Manhattan plot.
}
\description{
Manhattan plot
}
