% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.r
\name{dmrff.meta}
\alias{dmrff.meta}
\title{dmrff.meta}
\usage{
dmrff.meta(objects, maxgap = 500, p.cutoff = 0.05, verbose = T)
}
\arguments{
\item{objects}{List of objects generated by \code{\link{dmrff.pre}}
for each participating EWAS.}

\item{maxgap}{Maximum distance between consecutive features (Default: 500bp).}

\item{p.cutoff}{Unadjusted p-value cutoff for membership in a candidate DMR
(Default: 0.05).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}
}
\value{
A data frame listing all candidate regions and their summary statistics.
}
\description{
Identify differentially methylated regions by meta-analysing multiple studies
using variance-weighted fixed effects meta-analysis.
}
\examples{
pre1 <- dmrff.pre(est1, se1, meth1, ...)
pre2 <- dmrff.pre(est2, se2, meth2, ...)
...
pre9 <- dmrff.pre(est9, se9, meth9, ...)
meta <- dmrff.meta(list(pre1, pre2, ..., pre9))
meta$dmrs[which(meta$dmrs$p.adjust < 0.05 & meta$dmrs$n > 1), ]

}
