% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{dmrff.plot}
\alias{dmrff.plot}
\title{Plot a region}
\usage{
dmrff.plot(
  region.chr,
  region.start,
  region.end,
  estimate,
  se,
  chr,
  pos,
  expand = 1,
  ci = T,
  verbose = T
)
}
\arguments{
\item{estimate}{Vector of EWAS effect estimates
(corresponds to rows of \code{methylation}).}

\item{se}{Vector of standard errors of the coefficients.}

\item{chr}{Feature chromosome (corresponds to rows of \code{methylation}).}

\item{pos}{Feature chromosome position.}

\item{expand}{Proportion to expand to include sites beyond the beginning and end of the region (default: 1, i.e. plot 3x the length of the region).}

\item{ci}{Show confidence intervals rather than standard errors (default: TRUE).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}

\item{region.chr/start/end}{Genomic region to plot.}

\item{methylation}{Methylation matrix (rows=features, columns=samples).}
}
\value{
\code{\link{ggplot}} object showing the plot.
}
\description{
dmrff.plot
}
\details{
Calculate statistics for a set of genomic regions.
}
