% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.r
\name{dmrff.pre}
\alias{dmrff.pre}
\title{dmrff.pre}
\usage{
dmrff.pre(estimate, se, methylation, chr, pos, maxsize = 20,
  verbose = T)
}
\arguments{
\item{estimate}{Vector of association effect estimates
(corresponds to rows of \code{methylation}).}

\item{se}{Vector of standard errors of the effect estimates.}

\item{methylation}{Methylation matrix (rows=features, columns=samples).}

\item{chr}{Feature chromosome (corresponds to rows of \code{methylation}).}

\item{pos}{Feature chromosome position.}

\item{maxsize}{Maximum number of CpG sites in a DMR.}

\item{verbose}{If \code{TRUE} (default), then output status messages.}
}
\value{
A list object containing all information needed for inclusion in a meta-analysis.
}
\description{
Construct an object for including this dataset in a DMR meta-analysis.
}
