% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrff.r
\name{dmrff}
\alias{dmrff}
\title{dmrff}
\usage{
dmrff(
  estimate,
  se,
  p.value,
  methylation,
  chr,
  pos,
  maxgap = 500,
  p.cutoff = 0.05,
  minmem = T,
  verbose = T
)
}
\arguments{
\item{estimate}{Vector of association effect estimates
(corresponds to rows of \code{methylation}).}

\item{se}{Vector of standard errors of the effect estimates.}

\item{p.value}{Vector of p-values.}

\item{methylation}{Methylation matrix (rows=features, columns=samples).}

\item{chr}{Feature chromosome (corresponds to rows of \code{methylation}).}

\item{pos}{Feature chromosome position.}

\item{maxgap}{Maximum distance between consecutive features (Default: 500bp).}

\item{p.cutoff}{Unadjusted p-value cutoff for membership in a candidate DMR
(Default: 0.05).}

\item{verbose}{If \code{TRUE} (default), then output status messages.}
}
\value{
A data frame listing all candidate regions and their summary statistics.
}
\description{
Identifying differentially methylated regions efficiently with power and control.
}
\details{
Warning! Ensure that the order of the CpG sites corresponding to the the rows of `methylation`
match the order of the CpG sites corresponding to the other variables,
e.g. `estimate` and `chr`.
}
\examples{

dmrs <- dmrff(estimate, ## effect estimate for each CpG site
              se,       ## standard error of the estimate for each CpG site
              p.value,  ## p-value
              methylation, ## methylation matrix
              chr,      ## chromosome of each CpG site
              pos)      ## position of each CpG site
                     
dmrs[which(dmrs$p.adjust < 0.05 & dmrs$n > 1),]

}
