% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.r
\name{dmrff.sites}
\alias{dmrff.sites}
\title{dmrff.sites}
\usage{
dmrff.sites(regions, chr, pos)
}
\arguments{
\item{regions}{Data frame of genomic regions providing
chromosome (chr), start and end coordinates.}

\item{chr}{A vector providing the chromosome of each location.}

\item{pos}{A vector providing the chromosomal position of each location.}
}
\value{
A data frame identifying the region containing each genomic location.
}
\description{
Link genomic locations to a set of genomic regions.
}
\examples{

dmrs <- dmrff(estimate, se, p.value, methylation, chr, pos)
dmrs <- dmrs[which(dmrs$p.adjust < 0.05 & dmrs$n > 1),]
dmr.sites <- dmrff.sites(dmrs, chr, pos, estimate, se, p.value)
dmr.sites$dmr.z <- dmrs$z[dmr.sites$region]

}
